Enable the timer interrupt.
