+++ Kernel/swap.c
@@ -74,7 +74,7 @@ int swapread(uint16_t dev, blkno_t blkno, usize_t nbytes,
 #ifdef CONFIG_TRIM
 	while (nbytes != 0)
 	{
-		d_ioctl(dev, HDIO_TRIM, &blkno);
+		d_ioctl(dev, HDIO_TRIM, (void*)&blkno);
 		blkno++;
 		nbytes -= 512;
 	}
